[Executorch][threadpool] Add "private" method to reset num threads in threadpool
Pull Request resolved: https://github.com/pytorch/executorch/pull/2351

This diff introduces new API to reset threadpool. This will be short-lived
features. It is needed atm because we dont have a) a way to use only a subset of
threads to distribute work in the threadpool and b) API to create threadpool of
custom size and pass such a threadpool to backends like delegtes, cusotm ops,
optimized cpu op lib etc.
ghstack-source-id: 218361717
@exported-using-ghexport

Differential Revision: [D54766070](https://our.internmc.facebook.com/intern/diff/D54766070/)