test_ios comments on pybindings (#2324)
Summary:

Accoding to shoumikhin, this was a quick way to test the build for pybindings on macos, but pybindings isn't actually needed by the rest of the script.

Add comments to make it clear why these scripts are building the pybindings. Move the pybindings to the bottom so that nothing else can depend on them.

Reviewed By: shoumikhin

Differential Revision: D54695352